+++ modules/client-common/src/constants.ts
@@ -47,10 +47,10 @@ export const UNAVAILABLE_PROPOSAL_METADATA: ProposalMetadata = {
   resources: [],
 };
 
-const getGraphqlNode = (netowrk: SupportedNetwork): string => {
+const getGraphqlNode = (network: SupportedNetwork): string => {
   return `https://subgraph.satsuma-prod.com/qHR2wGfc5RLi6/aragon/osx-${
-    SupportedNetworksToGraphqlNetworks[netowrk]
-  }/version/v1.2.2/api`;
+    SupportedNetworksToGraphqlNetworks[network]
+  }/version/v1.3.0/api`;
 };
 
 export const GRAPHQL_NODES: { [K in SupportedNetwork]: { url: string }[] } = {