+++ modules/client/src/tokenVoting/internal/interfaces.ts
@@ -33,6 +33,7 @@ import {
   WrapTokensParams,
   WrapTokensStepValue,
 } from "../types";
+import { TokenVotingTokenCompatibility } from "./types";
 
 // TokenVoting
 
@@ -80,6 +81,7 @@ export interface ITokenVotingClientMethods {
     tokenAddress: string,
   ) => AsyncGenerator<UndelegateTokensStepValue>;
   getDelegatee: (tokenAddress: string) => Promise<string | null>;
+  isTokenVotingCompatibleToken: (tokenAddress: string) => Promise<TokenVotingTokenCompatibility>;
 }
 
 export interface ITokenVotingClientEncoding {