+++ modules/common/src/errors.ts
@@ -276,3 +276,12 @@ export class InvalidParameter extends SdkError {
     );
   }
 }
+export class NotAContractError extends SdkError {
+  constructor(cause?: Error) {
+    super(
+      "The provided address does not point to a contract",
+      cause,
+    );
+  }
+}
+